+++ scoreboard/src/main/java/net/apartium/cocoabeans/scoreboard/ScoreboardAction.java
@@ -1,0 +1,6 @@
+package net.apartium.cocoabeans.scoreboard;
+
+public enum ScoreboardAction {
+    CREATE_OR_UPDATE,
+    REMOVE
+}